move search logic to dropdown component
