+++ src/App.scss
@@ -1,7 +1,3 @@
 iframe {
   display: none;
 }
-
-
-
-
