+++ src/App.tsx
@@ -1,60 +1,16 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useState } from 'react';
 import './App.scss';
-import { debounce } from 'lodash';
+
 import { peopleFromServer } from './data/people';
 import { Person } from './types/Person';
 import { Dropdown } from './components/Dropdown';
 import { PersonInfo } from './components/PersonInfo';
 import { Placeholder } from './components/Placeholder';
 
-const DELAY = 500;
+const DELAY = 1000;
 
 export const App: React.FC = () => {
-  const [query, setQuery] = useState('');
-  const [debouncedQuery, setDebouncedQuery] = useState('');
   const [selectedPerson, setSelectedPerson] = useState<Person | null>(null);
-
-  const applyDebouncedQuery = useCallback(
-    debounce(setDebouncedQuery, DELAY),
-    [],
-  );
-
-  const filteredPeople = useMemo(() => (
-    peopleFromServer.filter(
-      ({ name }) => name.toLowerCase().includes(debouncedQuery.toLowerCase()),
-    )
-  ), [debouncedQuery]);
-
-  const visiblePeople: Person[] = useMemo(() => (
-    filteredPeople
-  ), [filteredPeople]);
-
-  const handleQueryChange = useCallback(
-    (newQuery: string) => {
-      setSelectedPerson(null);
-      setQuery(newQuery);
-      applyDebouncedQuery(newQuery.trim());
-    },
-    [applyDebouncedQuery],
-  );
-
-  const handlePersonClick = useCallback(
-    (person: Person) => {
-      setSelectedPerson(person);
-      setQuery(person.name);
-      setDebouncedQuery('');
-    },
-    [],
-  );
-
-  const handleResetQuery = useCallback(
-    () => {
-      setSelectedPerson(null);
-      setQuery('');
-      setDebouncedQuery('');
-    },
-    [],
-  );
 
   return (
     <main className="section">
@@ -65,14 +21,11 @@ export const App: React.FC = () => {
       }
 
       <Dropdown
-        people={visiblePeople}
-        query={query}
-        debouncedQuery={debouncedQuery}
-        onQueryChange={handleQueryChange}
-        onPersonClick={handlePersonClick}
-        onResetQuery={handleResetQuery}
-      />
+        people={peopleFromServer}
+        setSelectedPerson={setSelectedPerson}
+        delay={DELAY}
 
+      />
     </main>
   );
 };
