+++ src/components/DropdownItem/DropdownItem.scss
@@ -1,3 +1,3 @@
-dropdown-item:hover {
-    background-color: lightblue
-  }
+.dropdown-item:hover {
+  background-color: lightblue;
+}
